# swagger.yaml
...
openapi: 3.0.0
info:
  title: My Blog Application using Nodejs
  description: This blog application allows users to register, login, manage their profiles, create and manage posts, and comment on posts. It is built using Node.js with a RESTful API structure. The following sections provide details on the different API endpoints, their purposes, request formats, and expected responses.
  version: 1.0.0
paths:
...
  /comments:
    post:
      tags:
        - "Comments"
      summary: Adds Comment to a post
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/addComment"
      responses:
        "200":
          description: A successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/addComment"

        "404":
          description: Comment not found

    put:
      tags:
        - "Comments"
      summary: Edits a user's comment on a post
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/editComment"
      responses:
        "200":
          description: A successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/editComment"

        "404":
          description: Comment not found

    delete:
      tags:
        - "Comments"
      summary: Deletes a user's comment from a post
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/deleteComment"
      responses:
        "200":
          description: A successful response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/deleteComment"

        "404":
          description: Comment not found
components:
  schemas:
...
    addComment:
      type: object
      properties:
        comment:
          type: string
          example: Woowww you are doing amazing
        postId:
          type: string
          example: 66c393ed533981cfb9d860d0
        commentorsId:
          type: string
          example: 66c8fd1bcdc777f74c5e4691
    editComment:
      type: object
      properties:
        id:
          type: string
          example: 66be87f315e69f48c315416a
        comment:
          type: string
          example: This comment is to be edited
    deleteComment:
      type: object
      properties:
        id:
          type: string
          example: 66be87f315e69f48c315416a